Refactor functionality to locations under new mod struct
