+++ skbase/base/_meta.py
@@ -10,6 +10,7 @@ from inspect import isclass
 from typing import List
 
 from skbase.base._base import BaseEstimator
+from skbase.utils._nested_seq import flatten, is_flat, unflatten
 
 __author__: List[str] = ["mloning", "fkiraly"]
 __all__: List[str] = ["BaseMetaEstimator"]
@@ -587,76 +588,6 @@ class BaseMetaEstimator(BaseEstimator):
             self.set_tags(**{mid_tag_name: mid_tag_val})
         else:
             self.set_tags(**{mid_tag_name: mid_tag_val_not})
-
-
-def flatten(obj):
-    """Flatten nested list/tuple structure.
-
-    Parameters
-    ----------
-    obj: nested list/tuple structure
-
-    Returns
-    -------
-    list or tuple, tuple if obj was tuple, list otherwise
-        flat iterable, containing non-list/tuple elements in obj in same order as in obj
-
-    Example
-    -------
-    >>> flatten([1, 2, [3, (4, 5)], 6])
-    [1, 2, 3, 4, 5, 6]
-    """
-    if not isinstance(obj, (list, tuple)):
-        return [obj]
-    else:
-        return type(obj)([y for x in obj for y in flatten(x)])
-
-
-def unflatten(obj, template):
-    """Invert flattening, given template for nested list/tuple structure.
-
-    Parameters
-    ----------
-    obj : list or tuple of elements
-    template : nested list/tuple structure
-        number of non-list/tuple elements of obj and template must be equal
-
-    Returns
-    -------
-    rest : list or tuple of elements
-        has element bracketing exactly as `template`
-            and elements in sequence exactly as `obj`
-
-    Example
-    -------
-    >>> unflatten([1, 2, 3, 4, 5, 6], [6, 3, [5, (2, 4)], 1])
-    [1, 2, [3, (4, 5)], 6]
-    """
-    if not isinstance(template, (list, tuple)):
-        return obj[0]
-
-    list_or_tuple = type(template)
-    ls = [unflat_len(x) for x in template]
-    for i in range(1, len(ls)):
-        ls[i] += ls[i - 1]
-    ls = [0] + ls
-
-    res = [unflatten(obj[ls[i] : ls[i + 1]], template[i]) for i in range(len(ls) - 1)]
-
-    return list_or_tuple(res)
-
-
-def unflat_len(obj):
-    """Return number of non-list/tuple elements in obj."""
-    if not isinstance(obj, (list, tuple)):
-        return 1
-    else:
-        return sum([unflat_len(x) for x in obj])
-
-
-def is_flat(obj):
-    """Check whether list or tuple is flat, returns true if yes, false if nested."""
-    return not any(isinstance(x, (list, tuple)) for x in obj)
 
 
 class _HeterogenousMetaEstimator(BaseMetaEstimator):
